virtual_devices.py: Prepare for Config persistence and GUI
Extend the interface and fix bugs to enable persistent configurations and cumstom config dialogs.
Changed the implementation of VirtualInputDevice: The complete config is now stored in a dictionary to simplify the communication with the GUI.

The DeviceTypeCollection does not expect a dictionary of devices anymore but a List of devices.
The mapping from the name strings to the devices is done afterwards.
The devices' names must be unique!